… CMakeLists.txt
… README.md
… apps/sandbox_data/meshes/cube.obj
… apps/sandbox_data/meshes/quad_xy.obj
… src/CMakeLists.txt
… tests/tria/asset/mesh_obj_test.cpp