This pr switches to a custom obj parser (instead of using [tinyobjloader](https://github.com/tinyobjloader/tinyobjloader)). Main reason is performance, our obj parser takes roughly half the time. 

Also included in this pr is a new implementation for the index buffer generation which results in another 2x speed.

Timings for my machine:
```
Before debug:
* bunny.obj: 1.1 sec
* dragon.obj: 6.1 sec
Before release:
* bunny.obj: 334 ms
* dragon.obj: 1.6 sec

After debug:
* bunny.obj: 226 ms
* dragon.obj: 1.3 sec
After release:
* bunny.obj: 69 ms
* dragon.obj: 353 ms
```